refactor(bundler): move resolve_relative_import_with_context to import_transformer
- Moved resolve_relative_import_with_context from bundler.rs to import_transformer.rs as a standalone function
- Removed the wrapper function resolve_relative_import from bundler.rs
- Updated all callers to use the function from import_transformer module
- Function now takes entry_path and bundled_modules as parameters instead of accessing through self

This refactoring is part of the effort to split bundler.rs into smaller, more manageable modules.
