+++ crates/cribo/src/code_generator/module_transformer.rs
@@ -26,7 +26,10 @@ use crate::{
         context::ModuleTransformContext,
         globals::{GlobalsLifter, transform_globals_in_stmt},
         import_deduplicator,
-        import_transformer::{RecursiveImportTransformer, RecursiveImportTransformerParams},
+        import_transformer::{
+            RecursiveImportTransformer, RecursiveImportTransformerParams,
+            resolve_relative_import_with_context,
+        },
     },
     types::{FxIndexMap, FxIndexSet},
 };
@@ -80,10 +83,12 @@ pub fn transform_module_to_init_function<'a>(
     for stmt in &ast.body {
         if let Stmt::ImportFrom(import_from) = stmt {
             // Resolve the module to check if it's inlined
-            let resolved_module = bundler.resolve_relative_import_with_context(
+            let resolved_module = resolve_relative_import_with_context(
                 import_from,
                 ctx.module_name,
                 Some(ctx.module_path),
+                bundler.entry_path.as_deref(),
+                &bundler.bundled_modules,
             );
 
             if let Some(ref module) = resolved_module {
